+++ src/main/kotlin/com/github/etkachev/nxwebstorm/actionlisteners/DryRunAction.kt
@@ -1,7 +1,7 @@
 package com.github.etkachev.nxwebstorm.actionlisteners
 
-import com.github.etkachev.nxwebstorm.models.FormValueMap
-import com.github.etkachev.nxwebstorm.models.SchematicTypeEnum
+import com.github.etkachev.nxwebstorm.models.DryRunButtonData
+import com.github.etkachev.nxwebstorm.models.SchematicCommandData
 import com.github.etkachev.nxwebstorm.services.MyProjectService
 import com.github.etkachev.nxwebstorm.ui.RunTerminalWindow
 import com.github.etkachev.nxwebstorm.utils.getSchematicCommandFromValues
@@ -12,19 +12,17 @@ import javax.swing.AbstractAction
 
 class DryRunAction(
   proj: Project,
-  private val collection: String,
-  private val id: String,
-  private val formValues: FormValueMap,
-  private val dialog: DialogWrapper,
-  private val type: SchematicTypeEnum,
-  private val collectionPath: String
+  private val actionData: DryRunButtonData,
+  private val dialog: DialogWrapper
 ) : AbstractAction("Dry Run") {
   private var terminal = RunTerminalWindow(proj, "Dry Run")
   private var nxService = MyProjectService.getInstance(proj)
   override fun actionPerformed(e: ActionEvent?) {
+    val (collection, id, formValues, type, collectionPath) = actionData
     val values = formValues.formVal
     val projectType = nxService.nxProjectType
-    val command = getSchematicCommandFromValues(collection, id, values, projectType, true, type, collectionPath)
+    val schematicCommandData = SchematicCommandData(projectType, type, collectionPath)
+    val command = getSchematicCommandFromValues(collection, id, values, schematicCommandData, true)
     dialog.close(DialogWrapper.CANCEL_EXIT_CODE)
     terminal.runAndShow(command)
   }