+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Changed
 - Updated Readme to include the latest updates.
+- Renamed the column "Type" to "Collection" within the schematic table search.
 
 ### Deprecated